Fix first-open fork conversation menu availability
