+++ cmuxTests/WorkspaceForkConversationContextMenuTests.swift
@@ -4539,14 +4539,20 @@ struct WorkspaceForkConversationContextMenuTests {
             dateProvider: { Date(timeIntervalSince1970: 42) }
         )
 
-        await liveAgentIndex.refreshForkAvailabilityNow(
-            workspaceId: workspace.id,
-            panelId: panelId,
-            fallbackSnapshot: snapshotWithExecutable
+        #expect(
+            workspace.forkAgentConversationContextMenuPresentationAvailability(
+                forPanelId: panelId,
+                liveAgentIndex: liveAgentIndex
+            ) == .agentIndexRefreshing
+        )
+
+        await workspace.resolveForkAgentConversationContextMenuAvailability(
+            forPanelId: panelId,
+            liveAgentIndex: liveAgentIndex
         )
 
         #expect(
-            workspace.forkAgentConversationContextMenuOpenAvailability(
+            workspace.forkAgentConversationContextMenuPresentationAvailability(
                 forPanelId: panelId,
                 liveAgentIndex: liveAgentIndex
             ) == .available
